fully test and protect macros from `prelude` assumption
