+++ extendr-api/src/ownership.rs
@@ -270,6 +270,7 @@ impl Ownership {
 #[cfg(test)]
 mod test {
     use super::*;
+    use crate as extendr_api;
     use crate::*;
     use libR_sys::{Rf_ScalarInteger, Rf_protect};
