+++ extendr-api/src/wrapper/complexes.rs
@@ -91,6 +91,7 @@ impl TryFrom<Vec<c64>> for Complexes {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate as extendr_api;
 
     #[test]
     fn test_try_from_vec_c64_conversion() {
